--- src/main/java/Card/CreditCard.java
@@ -1,19 +1,0 @@
-package Card;
-
-public class CreditCard extends Card {
-    // Только для основного счёта.
-    public CreditCard(String name, double balance) {
-        super(name, balance);
-    }
-
-    // Одобряем кредит
-    @Override
-    public double withDraw(double min) {
-        double credit = balance - min;
-        if (min <= balance)
-            System.out.println("Вы сняли сумму в " + min + " " + "сумма на счету " + credit);
-        else
-            System.out.println("Вы сняли сумму в " + min + " " + ". Внимание!!! Вы взяли кредит. Задолженность на счету : " + credit);
-        return credit;
-    }
-}
