+++ src/main/java/start/Main.java
@@ -1,4 +1,6 @@
-import Menu.Menu;
+package start;
+
+import menu.Menu;
 
 import java.util.Scanner;
 
@@ -12,17 +14,16 @@ public class Main {
         Menu menu = new Menu();
         Scanner sc = new Scanner(System.in);
 
-
         System.out.println("Добро пожаловать в банк \"ПоПяццот\"");
         System.out.println("Выберете ваш счёт : ");
         System.out.println("1.Основной счёт \"Жирный кот\"");
         System.out.println("2.Подарочный счёт \"Ашчушчэне\"");
         int number = sc.nextInt();
         switch (number) {
-            case 1: // Вызов меню основного счёта / Доступен кредит
+            case 1:
                 menu.menuCard();
                 break;
-            case 2: // Вызов меню подарочного счёта / Доступен депозит
+            case 2:
                 menu.menuPresent();
                 break;
         }
