+++ src/test/java/card/CardTest.java
@@ -1,0 +1,25 @@
+//package card;
+//
+//import org.junit.Assert;
+//import org.junit.Before;
+//import org.junit.jupiter.api.Assertions;
+//import org.junit.jupiter.api.BeforeEach;
+//import org.junit.jupiter.api.Test;
+//
+//public class CardTest {
+//    private final double DEFAULT_BALANCE = 650.60 ;
+//    private final String DEFAULT_NAME = "MR.Neelixs" ;
+//    private final double delta = 0.00001;
+//    private Card card;
+//
+//    @BeforeEach
+//    public void setUP(){
+//        Card card = new Card(DEFAULT_NAME,DEFAULT_BALANCE);
+//    }
+//
+//    @Test
+//    public void printBalanceTest() {
+//        Assertions.assertEquals(DEFAULT_BALANCE,card.getBalance(),delta);
+//    }
+//
+//}
